Add generated byte pair encoding merges and vocab file with 50,000 tokens
There are still a few issues I need to work out with UTF-8 string concatenation, but I'm almost certain that can be fixed in a post-processing pass.

This data was generated on my 8-core Ryzen 7 5700G over about 2 weeks of mostly-continuous compute. It's imperfect data since I ran it on a small portion of the dataset and ramped it up as I improved performance and the compression ratio increased, but I believe it's good enough to be used for the LLM.